gemini-tools-tutorial/pyproject.toml
[project]
name = "gemini-tools-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.3.0",
    "python-dotenv>=1.0.1",
]
